+++ tests/integ/modin/resample/test_resample_fillna.py
@@ -67,7 +67,6 @@ def test_resample_fill_ser(interval, agg_func):
     )
 
 
-# One extra query to convert index to native pandas for dataframe constructor
 @interval
 @agg_func
 @sql_count_checker(query_count=3, join_count=1)
@@ -112,7 +111,6 @@ def resample_ffill_ser_one_gap(agg_func):
     )
 
 
-# One extra query to convert index to native pandas for dataframe constructor
 @interval
 @agg_func
 @sql_count_checker(query_count=3, join_count=1)
@@ -162,7 +160,6 @@ def test_resample_ffill_ser_missing_in_middle(interval, agg_func):
     )
 
 
-# One extra query to convert index to native pandas for dataframe constructor
 @interval
 @agg_func
 @sql_count_checker(query_count=3, join_count=1)
@@ -187,7 +184,6 @@ def test_resample_ffill_ffilled_with_none(interval, agg_func):
     )
 
 
-# One extra query to convert index to native pandas for dataframe constructor
 @interval
 @agg_func
 @sql_count_checker(query_count=3, join_count=1)